HierarchyObject support on interface bus.
The following change introduces changes to bus.py infrastructure that
enables support of struct interfaces.

The desing is done in a way that HierarchyObject means creation of new
Bus object. Therefore, the bus code builds a structure of busses with
two nodes - (i) ModifiableObject instance and (ii) Bus instance.
User is able to access the sub-bus using the dot notation.

The patch also provides a test case for Verilator using the simple
memory-based design.